requirements.txt
azure-storage-blob==12.20.0
azure-ai-ml==1.16.0
azure-identity==1.16.0
python-dotenv==1.0.1
Flask==3.0.3